Move defaults for ghg to utils
